minor refactor: create file for badge class
