+++ lib/leaderboard.rb
@@ -3,6 +3,7 @@ require 'faraday-http-cache'
 require 'json'
 require 'octokit'
 require 'open-uri'
+# require_relative 'badge'
 # require_relative 'member'
 
 include Concurrent
@@ -11,30 +12,6 @@ BASE_REPOS_URL = "#{BASE_API_URL}/repos".freeze
 ORG_REPOS_URL = "#{BASE_REPOS_URL}/ourtigarage".freeze
 SNAKE_URL = "#{ORG_REPOS_URL}/web-snake".freeze
 LEADERBOARD_URL = "#{ORG_REPOS_URL}/hacktoberfest-leaderboard".freeze
-
-# Class reprensenting a badge a player can earn
-class Badge
-  attr_reader :short, :title, :description
-
-  def initialize(short, title, description, &block)
-    @short = short
-    @title = title
-    @description = description
-    @block = block
-  end
-
-  def earned_by?(player)
-    # Return false if no evaluation predicate has been provided
-    return false if !@block || @block.call(player) == false
-    @block.call(player) == true || @block.call(player) > 0
-  end
-
-  def times_earned_by(player)
-    # Return 0 if no evaluation predicate has been provided
-    return 0 if !@block || @block.call(player) == false
-    @block.call(player) == true ? 1 : @block.call(player)
-  end
-end
 
 # This is the list of all badges that can be earned
 # The block that define the badge's challenge should return either an integer or a boolean
