+++ server.rb
@@ -1,6 +1,7 @@
 require 'erubis'
 require 'json'
 require 'sinatra'
+require_relative 'lib/badge'
 require_relative 'lib/member'
 require_relative 'lib/leaderboard'
